+++ src/main/java/cn/enbug/shop/shop/good/GoodAdminService.java
@@ -5,6 +5,7 @@ import cn.enbug.shop.common.service.GoodService;
 import com.jfinal.kit.LogKit;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 
 /**
  * @author Yang Zhizhuang
@@ -16,7 +17,7 @@ public class GoodAdminService {
     public static final GoodAdminService ME = new GoodAdminService();
     private static final GoodService GOOD_SERVICE = GoodService.ME;
 
-    Ret add(String token, String ip, String name, String description, double price, String avator, int number) {
+    Ret add(String token, String ip, String name, String description, BigDecimal price, String avator, int number) {
         try {
             if (GOOD_SERVICE.insert(token, ip, name, description, price, avator, number)) {
                 return Ret.succeed();
